Moving generator related classes to TuistGenerator (#306)
### Short description 

Tuist is a command line tool that offers several useful features related to managing Xcode projects at scale. The generation features are worthy of bundling in a library to enable external tools to leverage it and build on top of it. 

### Solution 

The `TuistGenerator` library has been previously created and several changes have been incrementally made to allow seamlessly migrating all the generation related code to it.

### Test Plan 

- Ensure unit tests continue to pass
  `swift test`

- Ensure acceptance tests continue to pass
  `bundle exec rake features`

- Manually generate any of the projects within `fixtures`
   `tuist generate`
  
### Notes 

`TuistGenerator` is not yet stable - breaking changes will most likely be made.