Add support for passing Organization name to project (#1062)
Resolves https://github.com/tuist/tuist/issues/1063

### Short description 📝

New Files created by Xcode use the Organization name set in the project file for copyright notices. The organization was left blank before this PR.

### Solution 📦

It is now possible to optionally pass the organization name to either the `Project` initializer or more generally in `Config`. The latter will override the former if set.

### Implementation 👩‍💻👨‍💻

- [x] Add new `organizationName` argument to Project
- [x] Add new `organizationName` GenerationOptions case to pass to Config
- [x] Use project `organizationName` during pbxproj generation via attributes
- [x] Add Tests