Remove symbols from test input/output file lists
When a project has a scheme that runs multiple test targets that rely on
the `Embed Precompiled Frameworks` script then the new build system
fails because there are multiple commands that produce the same target
file.

This commit prevents adding symbol files to the input/output file lists
for test targets to prevent this error from occuring.

Fixes tuist/tuist#919