Make environment logging configurable (#1782)
### Short description  

Printing the ENV for every script phase really pollutes build logs. This PR allows users to disable that.

### Solution 

I've added an argument to `ExecutionAction` to control whether logging is enabled. I've defaulted it to `true` to preserve existing behavior, and to make this a non-breaking change.

Users can control the enablement / disablement via specifying the `.disableShowEnvironmentVarsInScriptPhases` generation option in `Config.swift`

### Implementation 

- Extended the `ExecutionAction` struct, and made the initializer default the new value to true.
- In the unit tests, left the pre-action alone, and set the post-action to use `false`.
- Additionally introduced `XCTUnwrap` to clean up the optionals, primarily so the optional Bool test didn't need to go through contortions.
- Introduced the `.disableShowEnvironmentVarsInScriptPhases` generation option on `Config.swift`
- Introduced `TargetActionEnvironmentMapper` to control the `showEnvVarsInLog` property