Fix hashing calculation of SettingsDictionary
SettingsContentHasher used to receive SettingsDictionary and generate a string based on the keys and values. However, it wasn’t sorted according to the keys, so every run could yield a different hash for the settings.
I also changed the implementation of `hash(_ settingsDictionary: SettingsDictionary) ` to match the naming, and return the hash of the settings (and not simply a string containing its contents).