Improve performance by removing unneeded Glob directory cache (#2318)
The PR removes a cache that was used to prevent the check of isDirectory when listing files using glob. Unfortunately it seems the cache gets barely hit (actually I was not able to get any hit at all?)
This speeds things up to 65% from the time spent running the glob. Using the generated fixture "2000_sources" I've got almost 20% improvement! 

```
    Fixture       : 2000_sources
    Runs          : 5
    Result
        - cold : 14,50s  vs  17,71s (⬇︎ -3,21s -18,12%)
        - warm : 13,44s  vs  16,79s (⬇︎ -3,35s -19,97%)
```