Add storekit configuration file to the project files
This step is both a nice to have (that is, it is handy to have the file included in Xcode files navigator), but also very helpful. The StoreKit configuration shouldn’t be added to any target, but at the same time it should be retrievable by Xcode. Project files is the perfect solution, and developers don’t need to manually include the file somewhere