Add models to handle RunAction options (#2524)
* Add models to handle RunAction options

* Add storekit configuration file to the scheme generation

* Add storekit configuration file to the project files

This step is both a nice to have (that is, it is handy to have the file included in Xcode files navigator), but also very helpful. The StoreKit configuration shouldn’t be added to any target, but at the same time it should be retrievable by Xcode. Project files is the perfect solution, and developers don’t need to manually include the file somewhere

* Add linting for storekit configuration existence

* Add unit tests

* Improve documentation

* Add documentation to the website

* Update changelog

* Change StoreKit configuration linting from warning to error

* Refactor RunActionOptions to improve APIs ergonomics

* Format and lint files