Support XCFrameworks with missing architectures in the Info.plist (#3095)
Handle the case where users have deleted a binary from
an xcframework to save binary size (e.g maccatalyst) but
left the reference in the Info.plist. Provide a warning
to let users know their xcframework is missing an
architecture.

This is typically a recoverable error as long as there is 1
valid architecture tuist can use.