Add JSON format to graph command
This allows the `graph` command to export the project graph into JSON format. Users can then parse the JSON data and perform operations such as changeset detection, detecting which Xcode project is affected if a dependency is affected.

For example, given a PR, we can detect which Swift packages are affected by the files changed in the PR. Assuming the repository contains multiple Xcode projects, we then need to figure out which Xcode project is affected by the changed Swift packages. Finally we can then run tests on these affected Xcode projects.