Add JSON format to graph command (#3617)
* Add JSON format to graph command

This allows the `graph` command to export the project graph into JSON format. Users can then parse the JSON data and perform operations such as changeset detection, detecting which Xcode project is affected if a dependency is affected.

For example, given a PR, we can detect which Swift packages are affected by the files changed in the PR. Assuming the repository contains multiple Xcode projects, we then need to figure out which Xcode project is affected by the changed Swift packages. Finally we can then run tests on these affected Xcode projects.

* Update changelog

* Update graph command documentation page

* Add attribution and admonition

* Update unit test to decode JSON and compare

* Add a separate output JSON model

* Update GraphService to use new output JSON model

* Support disableAutogeneratedSchemes generation option

* Update graph command description

* Address all the suggestions

1. Move output structures to separate files.
2. Each output structure has its own `from` factory function.
3. Use `bug` as error message.
4. Use better syntax sugar.
5. Remove default scheme generation from output.

* Include JSON schema in command documentation

* Add table of JSON schema

* Fix format

* Format JSON output

* Update error type

* Fix comment typos and move `from` methods into GraphService file

Co-authored-by: yi.wang <yi.wang@cloudkitchens.com>