Fix xcframework import when framework name is different from xcframework name (#4401)
Summary:

- xcframeworks can have a different name from the frameworks or libraries they host
- To ensure the correct binary name is extracted the underlying framework name is used rather the xcframework name

Details:

The reason behind using framework's name instead xcframework's name is because we can have a xcframework with a different name than the library that is compiled on it. 

For example, I can have 2 xcframeworks NaluPlayback and NaluPlaybackLite, both will have their product's name as NaluPlayback in order to import both of them in the same way `import NaluPlayback` but when generating the xcframework, as it uses the scheme's name to create the xcframework's name they are going to be deployed as NaluPlaybackLite.xcframework and NaluPlayback.xcframework. But the linked libraries inside both xcframework will be NaluPlayback.framework

If we use the current solution, we will have an error like:

```
NaluPlaybackLite.xcframework is missing architecture ios-arm64/NaluPlayback.framework/NaluPlaybackLite defined in the Info.plist
```