Strip newlines from master.key when decrypting
Resolves: #4781

When using Tuist with encrypted code signing certificates in GitHub
Actions and storing the private encryption key in a repository secret,
GitHub Actions is appending a newline to the end of the master key. This
causes Tuist to be unable to decrypt the code signing certificates
because it is using the newline as part of the private key. This change
modifies the `SigningCipher.masterKey(at:)` function to read the
master.key file as a text file, trim the newlines, and then to get the
raw bytes for the characters to be hashed into the encryption key.