Refactor SigningCipher.masterKey(at:)
I revised the implementation of `SigningCipher.masterKey(at:)` based on
PR review feedback for #4782. The implementation of `masterKey(at:)` and
`readMasterKey(at:)` are extremely similar. Since the goal of the PR is
to read the master.key file as a text file and trim the newlines, I
removed most of the contents of `masterKey(at:)` and relied on the
duplicate implementation in `readMasterKey(at:)`. This change
significantly simplifies the implementation of `masterKey(at:)` and
removes duplicate code.