Fix support for master.key with final newline (#4782)
* Strip newlines from master.key when decrypting

Resolves: #4781

When using Tuist with encrypted code signing certificates in GitHub
Actions and storing the private encryption key in a repository secret,
GitHub Actions is appending a newline to the end of the master key. This
causes Tuist to be unable to decrypt the code signing certificates
because it is using the newline as part of the private key. This change
modifies the `SigningCipher.masterKey(at:)` function to read the
master.key file as a text file, trim the newlines, and then to get the
raw bytes for the characters to be hashed into the encryption key.

* Refactor SigningCipher.masterKey(at:)

I revised the implementation of `SigningCipher.masterKey(at:)` based on
PR review feedback for #4782. The implementation of `masterKey(at:)` and
`readMasterKey(at:)` are extremely similar. Since the goal of the PR is
to read the master.key file as a text file and trim the newlines, I
removed most of the contents of `masterKey(at:)` and relied on the
duplicate implementation in `readMasterKey(at:)`. This change
significantly simplifies the implementation of `masterKey(at:)` and
removes duplicate code.