Replace `customLLDBInitFile` from `String` to tuist types for file path. (#5071)
* improving type of customLLDBInitFile

improving the types of the customLLDBInitFile property in two ways. first, changing the type from 'String' to 'Path' for the customLLDBInitFile property in ProjectDescription.RunAction. second, we're changing the type from 'String' to 'AbsolutePath' for the customLLDBInitFile property in TuistGraph.RunAction

* using resolve function for mapping customLLDBInitFile path

customLLDBInitFile can using both relative paths such as "$(SRCROOT)/path/to/lldbinit" and absolute paths such as "/path/to/lldbinit".
makes sense to use the resolve function for different path options.

* apply relative path for customLLDBInitFile

* apply relative path to customLLDBInitFile using target path

* fix for some lint violation

* Update Sources/TuistGenerator/Generator/SchemeDescriptorsGenerator.swift

Co-authored-by: Kas <kwridan@users.noreply.github.com>

* add customLLDBInitFile contents to fixtures/ios_app_with_custom_scheme

* fix test case for SchemeDescriptorsGeneratorTests.test_schemeLaunchAction()

---------

Co-authored-by: jay.choi <jay.choi@kakaocorp.com>
Co-authored-by: Kas <kwridan@users.noreply.github.com>