Multi-Platform support - Phase 1: Introduce `TuistGraph.Destination` and migrate `TuistGraph.Target` to use it (#5132)
This is phase 1 which introduces some of the necessary updates internally within Tuist without any new public APIs and minimal observed differences in generated projects

This part of the work focuses on refactoring `TuistGraph.Target` to enable compilation and deployment to multiple destinations.  

**Changes**

`Platform` and `DeploymentTarget` have migrated to be:
- Supported Destinations: a list of devices that the target can be deployed to.  supported platforms are derived from this list.  This splits `iPhone` and `iPad` into distinct destintions, as well as `macOS` and `macCatalyst`.  
- Minimum Deployment: a minimum deployment is a version specifying a minimum requirement for a given OS.

`legacyPlatform` has been introduced which attempts to mimic the previous `platform` property for various parts of the internals which check for exclusive platforms as part of their logic. Those will be gradually migrated over to a more appropriate API (e.g. using `destinations` directly).

`platformFilter` has been updated to be `platformFilters` to support having more than one filter based on the supported `destinations`.

**Known impact**

- Build settings related to destination management (e.g. `SUPPORTS_MACCATALYST`, `SUPPORTS_MAC_DESIGNED_FOR_IPHONE_IPAD`, `SUPPORTS_XR_DESIGNED_FOR_IPHONE_IPAD`) are now explicitly set by Tuist and can no longer be overrides in xcconfig files
   - This is necessary to ensure a predicate behaviour when specifying destinations
- In some cases redundant `platformFilters` may appear in the generated project files (this will be addressed once multi destinations public API is introduced)

**Test Plan**

- Verify all unit and acceptance tests pass
- Manually generate a few fixtures and inspect the git diff produced in the generated fixtures