Load the first system dependency that exists given our supported platforms
When linking and building multi-platform targets, we only need a reference from a single platform.  So long as the framework supports

e.g. WidgetKit.framework is only available on iOS, macOS, and watchOS and when linking a multiplatform target, it does not matter what sdk on disk is linked.

Given how xcode projects can be edited, it’s possible to end up linking against an SDK path that is not a supported platform.  A multiplatform iOS/macOS app could link against the watchOS copy of `WidgetKit` without issue.