Multi-Platform support: Phase 2 (#5369)
* Migrate commands to use `exclusivePlatform`

Currently they will throw an error for targets that arent single platform

# Conflicts:
#	Sources/TuistAutomation/Utilities/XcodeBuildDestination+Find.swift

* Compare to exclusive platform

This test case validates current `ProjectDescription.Target` mapping and it should only ever allow mapping to a single platform.

* Filter other targets by what platforms are supported.

* Cache by concatentated platform strings

* Load the first system dependency that exists given our supported platforms

When linking and building multi-platform targets, we only need a reference from a single platform.  So long as the framework supports

e.g. WidgetKit.framework is only available on iOS, macOS, and watchOS and when linking a multiplatform target, it does not matter what sdk on disk is linked.

Given how xcode projects can be edited, it’s possible to end up linking against an SDK path that is not a supported platform.  A multiplatform iOS/macOS app could link against the watchOS copy of `WidgetKit` without issue.

* Update to use the platform we’re iterating over.  This was missed in the initial migration

* Add per sdk settings overlay for multiplatform targets.

* Migrate `exclusivePlatform` to `TuistKit` as it is dedicated to platforms for services

* Exclude some settings from being used for multiplatform targets

These are not included in default xcode projects or are generated elsewhere.  We exclude them here to allow our `overlay` function to work adding distinct values for specific SDKs when they arent equal

Right now that is just `LD_RUNPATH_SEARCH_PATHS`

* Removing these now unneeded errors

* Fix Tests

* Move platform selection down into loadDependency method.

* Adding tests for overlay method

* Move settings filtering into DefaultSettingsProvider