Remove personal plan (#263)
I noticed we have two plans associated with accounts, `:personal` and
`:team`, and I don't think that's accurate because personal is not a
plan per se. If an account represents a personal account, and therefore,
we need to apply specific authorization, that's information that we can
extract from the association between an account and a user.