Deployment improvements (#287)
I've created three workflows for each of the environments that we have:

- **Production:** It deploys every morning at 11:00.
- **Canary:** It deploys for every commit landing on `main`. This gives
us a time-window of a 4-5 working hours to catch issues that might
arise.
- **Staging:** It is manually deployed by triggering the GitHub Actions
workflow directly from GitHub's UI.

Moreover, I've created a **release** workflow that triggers every Monday
at 11:00 similar to the one that we have in tuist/tuist does:
- Collects all the changes by filtering the PRs that include
`changelog:*` labels.
- Elaborate a changelog with the changes.
- Updates the version constant in the repository
- Creates a GitHub release with the CHANGELOG changes.
- Pushes the version to the GitHub registry and tags it with `latest`
and the `{version}`

The motivation for moving away from continuous deployment for our
enterprise images is that customers will have a nicer changelog to see
what changes have been introduced. Otherwise, they are deploying
blindlessly and if something causes issues on their end it'll be harder
to debug.