Support optional frameworks
I added a required field to the TargetDependency.framework and
TargetDependency.xcframework cases to enable defining optional
frameworks. Frameworks and xcframeworks can be optional, allowing them
to be referenced but not included in the application bundle. I added
this feature to support using the Reveal server framework
(https://revealapp.com) to be used in a project. Reveal needs to be
present in the Debug builds, but needs to not be included in
distribution builds.

Resolves tuist/tuist#5520