Remove legacy dashboard, fix sign in with email (#328)
Removing legacy dashboard. I'm keeping the code around for the time
being, but we should remove it eventually as well. We should tell Adidas
that this is going to happen but their projects will still work (for the
time being).

I also noticed sign in with email doesn't work because the
protect_from_forgery was called too late.