Disable the Stripe logic (#330)
We are evolving our subscription model to be tailored to each
organization that goes above a push/pull data quota. Because of it, all
the logic to automatically manage the subscription and keep it up to
date based on the number of seats is no longer necessary. This PR also
mitigates the [500
error](https://appsignal.com/tuist-cloud/sites/64cac968d2a5e4d0cc68485b/exceptions/incidents/121)
that users are experiencing trying to create an organization.