Support optional frameworks and xcframeworks (#5525)
* Support optional frameworks

I added a required field to the TargetDependency.framework and
TargetDependency.xcframework cases to enable defining optional
frameworks. Frameworks and xcframeworks can be optional, allowing them
to be referenced but not included in the application bundle. I added
this feature to support using the Reveal server framework
(https://revealapp.com) to be used in a project. Reveal needs to be
present in the Debug builds, but needs to not be included in
distribution builds.

Resolves tuist/tuist#5520

* Fix unit test failure

I fixed a failing unit test where the JSON decoder failed to decode JSON
due to the change that I made.

* Revise required framework flag

I changed the required flag that I introduced in TargetDependency for
.framework and .xcframework to use the new FrameworkStatus enum instead
of a boolean value. This conforms closer to how it works for the .sdk
case.

* Fix formatting using lint-fix