Introduce static typing with sorbet (#334)
I'm adding static typing with Sorbet. Note that [static
checks](https://sorbet.org/docs/static) are disabled for all the files
so we should enable them gradually as we work on different files. Once
we've typed all the codebase and feel confident about the system, we can
introduce runtime checks to catch issues that might have gone unnoticed
with the static analysis.
I extended our CI pipeline to run type-checking as part of a job.