Delete router constraints (#331)
I believe the `devise_for` inside constraints at the router is the
culprit of the authentication issues some users are seeing. My guess (I
couldn't reproduce it), is that `devise_for` is setting global state in
the `Warden` global configuration, and since `devise_for` now executes
every time a request reaches the server, the behavior is now dependent
on the order of the requests reaching the server.
To my surprise, I tried to reproduce the 401 issue that we were
receiving when using a project token to authenticate and I was not able
to reproduce it in canary.