Fix authentication in Safari (#333)
I noticed the authentication doesn't work when using Safari. It turns
out that the `{"mode": "no-cors"}` that we were using with `fetch` was
not yielding the expected results in Safari. This PR adds a proxy to
include a CORS header to allow requests to the `/auth` resource in
origin `http://127.0.0.1:4545`. That should make it work :)