Add ` /ready` GET route (#358)
Addresses https://github.com/tuist/cloud/issues/357

Deployment tools such as Kubernetes require services to have a `/ready`
endpoint that they can use to verify that the server started
successfully. I'm adding the endpoint in this PR, and taking the
opportunity to add types to the files that I touched.