Fix exists when an object was created but not fully uploaded (#391)
When an upload fails due to a connection timeout, the object might have
already been created in the S3 bucket _but_ with no data. The download
would then fail as it wouldn't be a valid archive.

The fix is to use `get_object` instead and check the size of the object
– if it's 0, the `exists` method in `cache_service` returns false.

### 🎩

- I used the acceptance test from this
[PR](https://github.com/tuist/tuist-cloud/pull/21)
- Reuploaded one of the frameworks with an empty file
- Ran the test -> it would fail without the changes from this PR
- After applying the changes, the `exists` method returned false for the
given framework and we would then successfully reupload the object.