Add support for cache_category param (#405)
Adding support for `cache_category` param for cache endpoints. This
allows us to differ between build and test artifacts.

If you pass `builds` as a `cache_category`, the artifact will be saved
to `our-bucket/your-org/your-project/builds/hash/...`. For `tests`, the
path will be `our-bucket/your-org/your-project/tests/hash/...`.

If you don't pass a `cache_category`, we default to `builds` to stay
compatible with older clients.