Add support for multi-part uploads (#402)
I'm adding new endpoints to support multi-part uploads. Multi-part
uploads eliminate the need for persisting a state that indicates whether
the file was successfully uploaded or not, and also contribute to
reliability because we can chunk the artifacts into smaller pieces and
the requests take less to execute thus less prone to connectivity errors
or timeouts.

**Note** that the need for exponential backoff in the client is still
needed.