Add Stripe billing (#455)
This PR adds a new job that runs hourly to update the upload and
download cache events.

When using a remote cache, we query the upload cache events count and if
it's close to a threshold, we add a warning (as long as the `plan` is
`nil`). If over a threshold, we throw an error. In both cases, we refer
the user to contacts at support@tuist.io to get a quote on the Tuist
Cloud plan.