Add support for R2 (#512)
The support was easier than expected thanks to the R3-compatible
interface that R2 provides.
I took the opportunity to refactor the code a bit, and remove all the
logic around projects providing their own bucket. The column is still in
the database, but I decided to have a separate PR just for that.