Connect with BetterStack (#522)
I gave BetterStack a try as a tool to monitor the uptime of our
services, and was quite pleased with the experience so I'm moving the
monitoring there as well as the escalation tools, so we'll stop using
Zenduty. The tool also supports sending logs, information that they use
to detect any potential anomalies in the services' behaviour.

This PR sets the logger up when the project is hosted by Tuist.