Improve traceability of requests (#530)
Debugging failures currently is a bit tricky since we are unable to
trace failing requests on the client using the server data (e.g. logs).
The client is including the `x-request-id` header but it's useless in
its current state. This PR addresses that by:
- Adjusting our AppSignal configuration to include that header in the
reports.
- Adding a middleware that includes that header in the responses so that
we can present it in the client (a PR will follow)

Moreover, I included the `APP_REVISION` environment variable, which the
App Signal Ruby client uses to associate reports to a particular version
of our Tuist Cloud backend. This will also help correlate errors with a
particular version of the software.