GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Update tuist workspace (refs/gitbutler/Update-tuist-workspace)
   - Project.swift
 - Better Tuist declaration (refs/gitbutler/Better-Tuist-declaration)
   - Project.swift
 - Update test workflow (refs/gitbutler/Update-test-workflow)
   - Tuist/ProjectDescriptionHelpers/Target+Helpers.swift
   - Workspace.swift
   - Tuist/ProjectDescriptionHelpers/Module.swift
   - Tests/TuistTestAcceptanceTests/TestAcceptanceTests.swift
   - Tests/TuistGeneratorAcceptanceTests/GenerateAcceptanceTests.swift
   - Tests/TuistGenerateAcceptanceTests/GenerateAcceptanceTests.swift
   - Tests/TuistBuildAcceptanceTests/BuildAcceptanceTests.swift
   - Tests/TuistAutomationAcceptanceTests/TestAcceptanceTests.swift
   - Tests/TuistAutomationAcceptanceTests/BuildAcceptanceTests.swift
   - Project.swift
   - .github/workflows/tuist.yml

Your previous branch was: refs/heads/main

The sha for that commit was: 0cc37394a0357514832ffa178288629dae17bb14

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch