Version releases (#538)
As we onboard on-premise customers, we must have a versioning scheme, a
regular release cadence, and a changelog that they can use to keep up
with our updates. This PR addresses that by:

- **Introducing the versioning scheme `{MAJOR}.YY.0M.0D`:** I decided to
follow [CalVer](https://calver.org/) so an instance of Tuist Cloud can
know when it was released. We can do the same on the CLI and include
that information in a header. The server can then correlate that
information and return a warning that the CLI can present to nudge
on-premise customers to update their server version. This is important
since we want them to be in a different version of Tuist Cloud. We can
define a time window of 1 month, and if we detect they are behind,
developers will see the warning. The **major component** is a tool to
indicate that we are introducing breaking changes. We can bump it once a
year and take the opportunity to delete deprecated APIs that have been
around, for example, the non-multipart version of the uploads. We could
also align the major releases of the server with the major releases of
the CLI.
- **Publishing the release notes as part of the package:** When
publishing the image, we can use the annotation
`org.opencontainers.image.description` to include the CHANGELOG as part
of the release.

Moreover, I just noticed that we can invite people only to a certain
package, which means we no longer need to give anyone access to
`tuist/cloud` and we can get rid of the `tuist/cloud-enterprise`
repository.