Uniquely identify OAuth-2 authenticated users using their ID on the OAuth2 platform (#541)
Solves https://github.com/tuist/cloud/issues/521

We used the user email on the OAuth2 platform to look up the user in the
database. However, this approach has a big caveat. If the user has
multiple emails, we might check the wrong one. This PR implements a more
reliable solution where we store the unique id on the OAuth2 provider.