Make the app single-process (#579)
I noticed the project's Fly configuration is set up with two processes:
one that runs Traefik, Rails, and Phoenix and the other that runs Que's
background jobs. On the on-premise side, this implies that we have to
tell them to run two processes in the infrastructure, adding some
configuration complexity on their end.

This complexity is something that we should absorb. How? Since we are
now vendoring a process manager,
[hivemind](https://github.com/DarthSim/hivemind), which is able to spawn
multiple process, we can leverage that tooling to include the Queu
process too. Note that eventually, we'll have a single process, the
Erlang runtime, which will contain both the app and the background job
processing.

The changes are currently deployed at
[https://cloud-staging.tuist.io/](https://cloud-staging.tuist.io/)