Add Elixir utilities to get the configuration from either an encrypted file or environment variables (#580)
### Context
We need the same logic in Elixir as in Ruby for reading the
configuration from environment variables with a fallback on
configuration encrypted in the repository.

### What
The reading of the configuration from environment variables is
straightforward. For the encrypted content, I tried to read the Rails
files, but found myself navigating into a rabbit hole understanding how
Rails does the decryption. In the end, I decided that it was not worth
the effort and created an encrypted file for Phoenix using [this Elixir
package](https://github.com/kieraneglin/encrypted_secrets_ex) that
mimics that behavior. The one drawback is that we must remember to make
changes across the Ruby and Elixir encrypted files.

### Next
After this one is merged, I'll focus on implementing a middleware that
takes care of authenticating either a user or a project using their
token in the `Authorization` header.