Check the rediness of the 3 processes for considering a deploying successful (#585)
By introducing a proxy to forward traffic to Rails and Phoenix, we broke
the Fly logic that ensures the readiness of the deployment. Because
Traefik always succeeds in running and returning responses, regardless
of the status of the Phoenix and Rails app, it's possible that they both
fail to initialize and therefore, are unable to respond to HTTP
requests, but the process doesn't abort (e.g. Puma keeps running), even
if it's unable to respond requests.

To fix this, I'm using Fly's HTTP `[[services.http_checks]]` option to
send a request that navigates through `Traefik > Rails > Phoenix`. For
it to be considered successful, it has to navigate through the 3 HTTP
servers.