Move the IPv6 configuration to the Fly.toml (#596)
Documentation: https://github.com/tuist/tuist/pull/6072

Fly
[sets](https://fly.io/docs/elixir/getting-started/#important-ipv6-settings)
some environment variables to enable IPv6 for Ecto, which is fine for
our setup with Supabase, which uses IPv6. However, on-premise customers
might not support IPv6 in the network where the database lives, so we
should make that an opt-in configuration with default to IPv4. This PR
does that.