[Breaking] Introduce licenses for on-premise users (#646)
I'm introducing the Elixir side of our licensing solution. The
implementation is done natively in Rust to allow obfuscating the keys at
compilation time, which is impossible in Elixir. A module
`TuistCloud.Native` serves as an interface to the native code. Note that
the license is just read at launch time and loaded in memory for the app
to consume at runtime.
The two moments when the validations happen are:
- When the server boots
- When requests are sent to the API

In the future we might want to extend this to the UI, and also start
using the list of features to enable/disable features based on the
features the license has access to.