Use environment-specific keys with App Signal (#654)
I noticed some Phoenix data shows up in the Rails app in App Signal and
vice versa. That's because we were using an organization-scope key with
Rails, and App Signal forwarded the data to the wrong application. I'm
adjusting the configuration on the Rails side of things to use
app-scoped keys. Note that Phoenix is already using app-specific keys.