Handle invalid project id (#650)
Resolves: https://github.com/tuist/cloud/issues/649

If a `projectId` is not a full project id and is missing a slash, we
would error with an internal server error. Instead, we should handle the
error and provide an error message that details what went wrong.

I am not 100 % sure if my approach to the error handling is the correct
one here, so let me know if you'd want to do it differently.