Support database connections without SSL (#670)
One of our on-premise users reported that they could not launch Tuist
Cloud because the service blew up trying to connect to the database. It
turns out that we were assuming the connection with the database would
be SSL, which is not the case for instance when you try to set up a
constellation of services using Docker Compose.

This PR addresses that, and also adjusts our `Dockerfile` to work with
architectures other than AMD. This required pulling and compiling
OpenSSL from the sources, instead of pulling an architecture-specific
binary.

I'll open a PR shortly in the documentation site.