Fix getting cache event (#676)
The `get_cache_event` would throw an error if two cache events of the
same `event_type` would have the same `name`. That's completely valid
and we should be getting the cache event based on `hash` instead (there
shouldn't be two cache upload events that have the same `hash`).

In hindsight, this was very wrong. In Rails, we were getting the first
`CacheEvent` but that wouldn't get the one that actually corresponds to
the binary we're trying to download most of the time.