Implement a foundation for sending analytic events to Posthog and Attio (#677)
I'm laying the foundation for sending events to Posthog and Attio (CRM).
The solution I'm implementing in this PR leverages Erlang's `telemetry`
API for broadcasting and subscribing to events and the gen server in the
case of Posthog for optimal usage of their API by sending events in
batches. The implementation is generic enough to easily add/remove
handlers as needed. For example:

```elixir
TuistCloud.Analytics.user_authenticate("pedro@tuist.io")
```