Write cache event update in a more performant query (#680)
Resolves: https://github.com/tuist/cloud/issues/674

We had to disable the update cache event count because of performance
reasons. The initial implementation was very suboptimal as we loaded
_all_ cache events in memory (and there can be a _huge_ amount of them).

Now, we map only over all accounts using Stream. This should improve
performance by a ton. I'm not sure if there's a good way for us to
benchmark this before shipping it.