Fix account creation failures due to account handle collisions (#721)
Since we obtained the account handle from the email, there's a chance
that there's another account with the same handle. This PR adds some
logic to handle that scenario by using suffixes to find a unique handle.
We used to have this logic in Rails, but we forgot to port it.