Improve performance of loading dashboard by leveraging Timescale (#719)
You can test this with the seed data that 10k events (and you can do
even more). Before, the load was _endless_ whereas now it's quite speedy
✨

For on-premise customers, they will need to install Timescale extension
themselves. When hosting a docker image, they should be able to follow
[these](https://docs.timescale.com/self-hosted/latest/install/installation-docker/)
instructions. I will add it to the docs.

I also tested things in the staging env.