Extend the analytics controller to take selective testing data, and add internal interfaces to fetch information for billing purposes (#734)
I'm extending the analytics controller to read selective testing
analytics from the client (PR to follow), and utilities to fetch the
usage of the current month, which we can use from the billing logic.
Note that I'm starting to model the platform's features with Phoenix
contexts. In essence, `TuistCloud.Cache` is for the Tuist Cache
functionality, and `TuistCloud.Tests` is for the Tuist Tests
functionality.