Fix failing migrations when booting the app with a clean database (#751)
When we transitioned from Rails, we dumped the existing database schema
into a `structure.sql` file that we need to dump in new databases before
running migrations. However, we forgot to load the schema when spinning
up new instances of Tuist Cloud (e.g., on-premise) with a clean
database, which causes the migration to fail. Note that we didn't run
into this issue because our database already has schema.
I've followed the steps [described
here](https://fly.io/phoenix-files/loading-structure-sql-on-prod-without-mix/)
to run the