Return forbidden when creating an analytics even from CI and authenticated as project (#745)
Resolves: https://github.com/tuist/app/issues/744

It seems some projects use Tuist Cloud with old Tuist versions (3.x). In
those versions, it was possible to use the project token in the CI
environment. However, when creating the command event, we made the
assumption that if `is_ci == false` then the authenticated subject must
be a user. We want to keep that assumption, but we should return a
`forbidden` error in those cases.

The project that triggered was `EtsyApp` but I assume it's some legacy
project or setup. When the request fails, the command just doesn't get
reported, so if the command event post fails, it shouldn't be a blocker
in 3.x versions.