Redirect to home if try to access an invalid auth path (#792)
Fixes: https://github.com/tuist/app/issues/766

If the server is hit with `/users/auth`, Ueberauth tries to invoke
`request` in the controller, which leads to a 5xx error because the
function is not defined. This PR provides an implementation for the
function that redirects the user to home with a flash message.

Why are people hitting those endpoints? It's still unknown to me. Some
requests are triggered by requests to `/users/auth/gitlab`, which is not
a path that you can navigate from any page in the app. So, I guess that
someone is trying to hit that URL by replacing the path component in the
browser (or maybe a bot?). Other requests are going to
`/users/auth/google_oauth2`, which is also unsupported because the
requests should go to `/users/auth/google`