Show test details (#794)
Counterpart Tuist client PR: https://github.com/tuist/tuist/pull/6373

This PR:
- Adds endpoints for uploading the invocation record and the individual
result bundle objects
- Adds a new card in the run detail for tests

To get the data for test summary, we download the result bundle objects
in memory and parse them on the server. We currently don't save anything
in the database as that's not needed for now. However, we will need to
save some data in the database in the future to unlock test flakiness
detection. That will come in a future PR.

For parsing the result bundles, I used
[this](https://github.com/davidahouse/XCResultKit/blob/main/XCResultFormat.md)
schema as an inspiration.

<img width="1270" alt="image"
src="https://github.com/tuist/app/assets/9371695/48f3d2e7-022b-4ffb-baa9-c5d7acda1887">